+++ sql/GeneraliDB/Tables/dbo.Attendance.sql
@@ -14,7 +14,7 @@ CREATE TABLE [dbo].[Attendance](
 	[ParentCategory] [nvarchar](100) NULL,
 	[SubCategory] [nvarchar](100) NULL,
 	[RecordDateTime] [datetime] NULL,
-PRIMARY KEY CLUSTERED 
+ CONSTRAINT [PK_Attendance] PRIMARY KEY CLUSTERED 
 (
 	[ID] ASC
 )WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]