+++ sql/GeneraliDB/Tables/dbo.CSVImportLog.sql
@@ -1,12 +1,12 @@
 ﻿USE [Generali]
 GO
-ALTER TABLE [dbo].[CSVImportLog] DROP CONSTRAINT [DF__CSVImport__Statu__1E6F845E]
+ALTER TABLE [dbo].[CSVImportLog] DROP CONSTRAINT [DF_CSVImportLog_Status]
 GO
-ALTER TABLE [dbo].[CSVImportLog] DROP CONSTRAINT [DF__CSVImport__RowsU__1D7B6025]
+ALTER TABLE [dbo].[CSVImportLog] DROP CONSTRAINT [DF_CSVImportLog_RowsUpdated]
 GO
-ALTER TABLE [dbo].[CSVImportLog] DROP CONSTRAINT [DF__CSVImport__RowsI__1C873BEC]
+ALTER TABLE [dbo].[CSVImportLog] DROP CONSTRAINT [DF_CSVImportLog_RowsInserted]
 GO
-ALTER TABLE [dbo].[CSVImportLog] DROP CONSTRAINT [DF__CSVImport__Start__1B9317B3]
+ALTER TABLE [dbo].[CSVImportLog] DROP CONSTRAINT [DF_CSVImportLog_StartedAt]
 GO
 DROP INDEX [IX_CSVImportLog_StartedAt] ON [dbo].[CSVImportLog]
 GO
@@ -27,7 +27,7 @@ CREATE TABLE [dbo].[CSVImportLog](
 	[MinScanDatum] [datetime] NULL,
 	[MaxScanDatum] [datetime] NULL,
 	[Status] [varchar](20) NOT NULL,
-PRIMARY KEY CLUSTERED 
+ CONSTRAINT [PK_CSVImportLog] PRIMARY KEY CLUSTERED 
 (
 	[ID] ASC
 )WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
@@ -38,11 +38,11 @@ CREATE NONCLUSTERED INDEX [IX_CSVImportLog_StartedAt] ON [dbo].[CSVImportLog]
 	[StartedAt] DESC
 )WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, SORT_IN_TEMPDB = OFF, DROP_EXISTING = OFF, ONLINE = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]
 GO
-ALTER TABLE [dbo].[CSVImportLog] ADD  DEFAULT (getdate()) FOR [StartedAt]
+ALTER TABLE [dbo].[CSVImportLog] ADD  CONSTRAINT [DF_CSVImportLog_StartedAt]  DEFAULT (getdate()) FOR [StartedAt]
 GO
-ALTER TABLE [dbo].[CSVImportLog] ADD  DEFAULT ((0)) FOR [RowsInserted]
+ALTER TABLE [dbo].[CSVImportLog] ADD  CONSTRAINT [DF_CSVImportLog_RowsInserted]  DEFAULT ((0)) FOR [RowsInserted]
 GO
-ALTER TABLE [dbo].[CSVImportLog] ADD  DEFAULT ((0)) FOR [RowsUpdated]
+ALTER TABLE [dbo].[CSVImportLog] ADD  CONSTRAINT [DF_CSVImportLog_RowsUpdated]  DEFAULT ((0)) FOR [RowsUpdated]
 GO
-ALTER TABLE [dbo].[CSVImportLog] ADD  DEFAULT ('running') FOR [Status]
+ALTER TABLE [dbo].[CSVImportLog] ADD  CONSTRAINT [DF_CSVImportLog_Status]  DEFAULT ('running') FOR [Status]
 GO