+++ sql/GeneraliDB/Tables/dbo.CategoryTranslation.sql
@@ -12,7 +12,7 @@ CREATE TABLE [dbo].[CategoryTranslation](
 	[OriginalValue] [nvarchar](255) NULL,
 	[Locale] [nvarchar](10) NULL,
 	[TranslatedValue] [nvarchar](255) NULL,
-PRIMARY KEY CLUSTERED 
+ CONSTRAINT [PK_CategoryTranslation] PRIMARY KEY CLUSTERED 
 (
 	[ID] ASC
 )WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON) ON [PRIMARY]